requirements.txt
astroid==2.3.0
beautifulsoup4==4.8.1
colorama==0.4.1
cx-Oracle==7.3.0
dj-database-url==0.5.0
Django==2.2.7
django-appconf==1.0.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-grappelli==2.13.2
django-heroku==0.3.1
django-select2==7.1.2
django-xhtml2pdf==0.0.4
gunicorn==20.0.4
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
mysql-connector-python==8.0.18
Pillow==6.2.1
protobuf==3.11.2
psycopg2==2.8.4
pylint==2.4.1
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.32
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.5
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.11.2
xhtml2pdf==0.2.3
django-extensions
python-decouple
redis
celery
crontab
django-solo
python-dateutil